requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==2.2
django-bootstrap4==3.0.1
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
django-tinymce==2.8.0